Refactor to silo apy chart component
