+++ src/components/charts/SeasonalAPYChart.tsx
@@ -2,38 +2,15 @@ import FrameAnimator from "@/components/LoadingSpinner.tsx";
 import { formatDate } from "@/utils/format";
 import { UseSeasonalResult } from "@/utils/types";
 import { cn } from "@/utils/utils";
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 import { CloseIconAlt } from "../Icons";
 import LineChart, { LineChartData } from "./LineChart";
-import {
-  metallicGreenStrokeGradientFn,
-  metallicMorningAreaGradientFn,
-  metallicMorningStrokeGradientFn,
-} from "./chartHelpers";
-import { TIME_TABS, TimeTab } from "./TimeTabs";
+import { metallicGreenStrokeGradientFn } from "./chartHelpers";
+import { SeasonalChartData } from "./SeasonalChart";
+import TimeTabsSelector, { TimeTab } from "./TimeTabs";
 
-export const tabToSeasonalLookback = (tab: TimeTab): number => {
-  if (tab === TimeTab.Week) {
-    return 24 * 7 - 1;
-  } else if (tab === TimeTab.Month) {
-    return 24 * 30 - 1;
-  }
-  return Number.MAX_SAFE_INTEGER;
-};
-
-export interface SeasonalChartData {
-  season: number;
-  value: number;
-  timestamp: Date;
-}
-
-// Interface for y-axis range configuration by time period
-export interface YAxisRangeConfig {
-  min?: number;
-  max?: number;
-}
-
-interface SeasonalChartProps {
+// TODO(pp): Does seasonal apy chart need any props at all?
+interface SeasonalAPYChartProps {
   title: string;
   size: "small" | "large";
   activeTab: TimeTab;
@@ -41,25 +18,13 @@ interface SeasonalChartProps {
   useSeasonalResult: UseSeasonalResult;
   valueFormatter: (value: number) => string;
   tickValueFormatter?: (value: number) => string;
-  fillArea?: boolean;
   statVariant?: "explorer" | "non-colored";
   className?: string;
-  useLogarithmicScale?: boolean;
-  showReferenceLineAtOne?: boolean;
-  // New props for custom y-axis ranges
-  yAxisRanges?: {
-    [TimeTab.Week]?: YAxisRangeConfig;
-    [TimeTab.Month]?: YAxisRangeConfig;
-    [TimeTab.AllTime]?: YAxisRangeConfig;
-  };
 }
 
-const morningStrokeGradients = [metallicMorningStrokeGradientFn];
 const greenStrokeGradients = [metallicGreenStrokeGradientFn];
 
-const areaGradients = [metallicMorningAreaGradientFn];
-
-const SeasonalChart = ({
+const SeasonalAPYChart = ({
   title,
   size,
   activeTab,
@@ -67,13 +32,9 @@ const SeasonalChart = ({
   useSeasonalResult,
   valueFormatter,
   tickValueFormatter,
-  fillArea,
   statVariant = "explorer",
   className,
-  useLogarithmicScale = false,
-  showReferenceLineAtOne = false,
-  yAxisRanges,
-}: SeasonalChartProps) => {
+}: SeasonalAPYChartProps) => {
   const [allData, setAllData] = useState<SeasonalChartData[] | null>(null);
   const [displayData, setDisplayData] = useState<SeasonalChartData | null>(null);
 
@@ -86,8 +47,8 @@ const SeasonalChart = ({
     }
   }, [inputData, allData]);
 
-  const handleChangeTabFactory = useCallback(
-    (tab: TimeTab) => () => {
+  const handleChangeTab = useCallback(
+    (tab: TimeTab) => {
       onChangeTab(tab);
       setAllData(null);
       setDisplayData(null);
@@ -98,32 +59,13 @@ const SeasonalChart = ({
   const chartData = useMemo<LineChartData[]>(() => {
     if (allData) {
       return allData.map((d) => ({
-        values: [useLogarithmicScale ? Math.max(0.000001, d.value) : d.value],
+        // Can't render 0 values on a log scale
+        values: [Math.max(0.000001, d.value)],
         timestamp: d.timestamp,
       }));
     }
     return [];
-  }, [allData, useLogarithmicScale]);
-
-  const horizontalReferenceLines = useMemo(() => {
-    if (showReferenceLineAtOne) {
-      return [
-        {
-          value: 1,
-          color: "#9C9C9C",
-          dash: [2, 10],
-          label: "$1.00 target",
-        },
-      ];
-    }
-    return [];
-  }, [showReferenceLineAtOne]);
-
-  // Get the current y-axis range based on active tab
-  const currentYAxisRange = useMemo(() => {
-    if (!yAxisRanges) return undefined;
-    return yAxisRanges[activeTab];
-  }, [yAxisRanges, activeTab]);
+  }, [allData]);
 
   const handleMouseOver = useCallback(
     (index: number) => {
@@ -142,18 +84,7 @@ const SeasonalChart = ({
         >
           {title}
         </div>
-        <div className="flex gap-4 sm:gap-8">
-          {TIME_TABS.map((tabName: string, idx: number) => (
-            <div
-              key={tabName}
-              data-state={activeTab === idx ? "active" : "inactive"}
-              onClick={handleChangeTabFactory(idx)}
-              className={`${activeTab === idx ? "text-pinto-green-3 sm:text-pinto-green-3" : "text-pinto-light sm:text-pinto-light"} pinto-sm sm:pinto-body-light cursor-pointer data-[state=inactive]:hover:text-pinto-green-4`}
-            >
-              {tabName}
-            </div>
-          ))}
-        </div>
+        <TimeTabsSelector tab={activeTab} setTab={handleChangeTab} />
       </div>
 
       {!allData && !displayData && (
@@ -177,19 +108,6 @@ const SeasonalChart = ({
             </div>
           </div>
         </>
-      )}
-      {allData && allData.length === 0 && !useSeasonalResult.isLoading && (
-        <div
-          className={`relative w-full flex items-center justify-center ${size === "small" ? "aspect-3/1" : "aspect-6/1"}`}
-          style={{
-            paddingBottom: `calc(85px + ${size === "small" ? "33.33%" : "16.67%"})`,
-            height: "0",
-          }}
-        >
-          <div className="absolute inset-0 flex items-center justify-center">
-            <div className="pinto-body-light">No silo interactions from connected wallet</div>
-          </div>
-        </div>
       )}
       {allData && displayData && (
         <>
@@ -219,14 +137,12 @@ const SeasonalChart = ({
                   data={chartData}
                   xKey="timestamp"
                   size={size}
-                  makeLineGradients={fillArea ? morningStrokeGradients : greenStrokeGradients}
-                  makeAreaGradients={fillArea ? areaGradients : undefined}
+                  makeLineGradients={greenStrokeGradients}
                   valueFormatter={tickValueFormatter}
                   onMouseOver={handleMouseOver}
-                  useLogarithmicScale={useLogarithmicScale}
-                  horizontalReferenceLines={horizontalReferenceLines}
-                  yAxisMin={currentYAxisRange?.min}
-                  yAxisMax={currentYAxisRange?.max}
+                  useLogarithmicScale={true}
+                  // yAxisMin={currentYAxisRange?.min}
+                  // yAxisMax={currentYAxisRange?.max}
                 />
               </div>
             )}
@@ -236,4 +152,4 @@ const SeasonalChart = ({
     </div>
   );
 };
-export default SeasonalChart;
+export default SeasonalAPYChart;
